requirements.txt
dotenv
openai
pandas
gymnasium
numpy
matplotlib
fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pymongo
motor
mongoengine
pdfplumber
